test: move field-base mock fields to test/fixtures
