+++ packages/field-base/test/visual/field-base.test.js
@@ -1,6 +1,6 @@
 import { fixtureSync } from '@vaadin/testing-helpers/dist/fixture.js';
 import { visualDiff } from '@web/test-runner-visual-regression';
-import './mock-field.js';
+import '../fixtures/mock-field.js';
 
 describe('field-base', () => {
   let div, element;
